+++ packages/admin/admin-stories/src/apollo-swapi-story.decorator.tsx
@@ -1,25 +1,27 @@
 import { ApolloClient, ApolloLink, ApolloProvider, createHttpLink, InMemoryCache } from "@apollo/client";
-import { createErrorDialogApolloLink, useErrorDialog } from "@comet/admin";
+import { createErrorDialogApolloLink } from "@comet/admin";
 import { LegacyStoryFn } from "@storybook/addons";
 import * as React from "react";
 
 import { DecoratorContext } from "./storyHelpers";
 
-const SwapiApolloProvider: React.FunctionComponent = ({ children }) => {
-    const errorDialog = useErrorDialog();
-
+const createApolloClient = () => {
     const link = ApolloLink.from([
-        createErrorDialogApolloLink({ errorDialog }),
+        createErrorDialogApolloLink(),
         createHttpLink({
             uri: `https://swapi-graphql.netlify.app/.netlify/functions/index`, // Test API here: https://graphql.org/swapi-graphql
         }),
     ]);
     const cache = new InMemoryCache();
-
-    const apolloClient = new ApolloClient({
+    return new ApolloClient({
         link,
         cache,
     });
+};
+
+const apolloClient = createApolloClient();
+
+const SwapiApolloProvider: React.FunctionComponent = ({ children }) => {
     return <ApolloProvider client={apolloClient}>{children}</ApolloProvider>;
 };