+++ packages/admin/admin-theme/src/componentsTheme/MuiAutocomplete.tsx
@@ -1,16 +1,19 @@
 import { Clear } from "@comet/admin-icons";
-import { Components } from "@mui/material/styles/components";
-import { Spacing } from "@mui/system";
 import * as React from "react";
 
-export const getMuiAutocomplete = (spacing: Spacing): Components["MuiAutocomplete"] => ({
+import { mergeOverrideStyles } from "../utils/mergeOverrideStyles";
+import { GetMuiComponentTheme } from "./getComponentsTheme";
+
+export const getMuiAutocomplete: GetMuiComponentTheme<"MuiAutocomplete"> = (component, { spacing }) => ({
+    ...component,
     defaultProps: {
         clearIcon: <Clear color="action" />,
+        ...component?.defaultProps,
     },
-    styleOverrides: {
+    styleOverrides: mergeOverrideStyles<"MuiAutocomplete">(component?.styleOverrides, {
         endAdornment: {
             top: "auto",
             right: spacing(2),
         },
-    },
+    }),
 });