+++ packages/admin/admin-theme/src/componentsTheme/MuiDialog.ts
@@ -1,8 +1,11 @@
 import { Components } from "@mui/material/styles/components";
-import { Spacing } from "@mui/system";
 
-export const getMuiDialog = (spacing: Spacing): Components["MuiDialog"] => ({
-    styleOverrides: {
+import { mergeOverrideStyles } from "../utils/mergeOverrideStyles";
+import { GetMuiComponentTheme } from "./getComponentsTheme";
+
+export const getMuiDialog: GetMuiComponentTheme<"MuiDialog"> = (component, { spacing }): Components["MuiDialog"] => ({
+    ...component,
+    styleOverrides: mergeOverrideStyles<"MuiDialog">(component?.styleOverrides, {
         paper: {
             borderRadius: 4,
             width: "100%",
@@ -29,5 +32,5 @@ export const getMuiDialog = (spacing: Spacing): Components["MuiDialog"] => ({
         paperFullWidth: {
             width: `calc(100% - ${spacing(16)})`,
         },
-    },
+    }),
 });