+++ packages/admin/admin-theme/src/componentsTheme/MuiTypography.ts
@@ -1,9 +1,11 @@
-import { Components } from "@mui/material/styles/components";
+import { mergeOverrideStyles } from "../utils/mergeOverrideStyles";
+import { GetMuiComponentTheme } from "./getComponentsTheme";
 
-export const getMuiTypography = (): Components["MuiTypography"] => ({
-    styleOverrides: {
+export const getMuiTypography: GetMuiComponentTheme<"MuiTypography"> = (component) => ({
+    ...component,
+    styleOverrides: mergeOverrideStyles<"MuiTypography">(component?.styleOverrides, {
         gutterBottom: {
             marginBottom: 20,
         },
-    },
+    }),
 });