+++ packages/admin/admin-theme/src/createCometTheme.ts
@@ -1,42 +1,42 @@
 import { createTheme, Theme, ThemeOptions } from "@mui/material";
-import { Palette, PaletteOptions } from "@mui/material/styles";
-import createPalette from "@mui/material/styles/createPalette";
-import createTypography, { Typography, TypographyOptions } from "@mui/material/styles/createTypography";
-import { createSpacing, Spacing, SpacingOptions } from "@mui/system";
-import merge from "lodash.merge";
+import createPalette, { PaletteOptions } from "@mui/material/styles/createPalette";
+import createTypography, { TypographyOptions } from "@mui/material/styles/createTypography";
+import { createSpacing } from "@mui/system";
+import { deepmerge } from "@mui/utils";
 
 import { getComponentsTheme } from "./componentsTheme/getComponentsTheme";
 import { paletteOptions as cometPaletteOptions } from "./paletteOptions";
 import { shadows } from "./shadows";
 import { typographyOptions as cometTypographyOptions } from "./typographyOptions";
 
-export const createCometTheme = (customThemeOptions: ThemeOptions | undefined = {}): Theme => {
-    const customPaletteOptions: PaletteOptions = customThemeOptions?.palette ? customThemeOptions.palette : {};
-    const paletteOptions: PaletteOptions = merge(cometPaletteOptions, customPaletteOptions);
-    const palette: Palette = createPalette(paletteOptions);
+export const createCometTheme = ({
+    palette: passedPaletteOptions = {},
+    typography: passedTypographyOptions = {},
+    spacing: passedSpacingOptions = 5,
+    components: passedComponentsOptions = {},
+    ...restPassedOptions
+}: ThemeOptions | undefined = {}): Theme => {
+    const paletteOptions: PaletteOptions = deepmerge<PaletteOptions>(cometPaletteOptions, passedPaletteOptions);
+    const palette = createPalette(paletteOptions);
 
-    const customTypographyOptions: TypographyOptions | ((palette: Palette) => TypographyOptions) = customThemeOptions?.typography
-        ? customThemeOptions.typography
-        : {};
-    const customTypographyOptionsObject: TypographyOptions =
-        typeof customTypographyOptions === "function" ? customTypographyOptions(palette) : customTypographyOptions;
-    const typographyOptions: TypographyOptions = merge(cometTypographyOptions, customTypographyOptionsObject);
-    const typography: Typography = createTypography(palette, typographyOptions);
+    const passedTypographyOptionsObject: TypographyOptions =
+        typeof passedTypographyOptions === "function" ? passedTypographyOptions(palette) : passedTypographyOptions;
+    const typographyOptions: TypographyOptions = deepmerge<TypographyOptions>(cometTypographyOptions, passedTypographyOptionsObject);
+    const typography = createTypography(palette, typographyOptions);
 
-    const spacingOptions: SpacingOptions = customThemeOptions?.spacing === undefined ? 5 : customThemeOptions.spacing;
-    const spacing: Spacing = createSpacing(spacingOptions);
+    const spacing = createSpacing(passedSpacingOptions);
 
     const cometThemeOptions: ThemeOptions = {
-        spacing: spacingOptions,
+        spacing: passedSpacingOptions,
         palette: paletteOptions,
         typography: typographyOptions,
         shape: {
             borderRadius: 2,
         },
         shadows,
-        components: getComponentsTheme(palette, typography, spacing),
+        components: getComponentsTheme(passedComponentsOptions, { palette, typography, spacing }),
     };
 
-    const themeOptions: ThemeOptions = merge(cometThemeOptions, customThemeOptions);
+    const themeOptions = deepmerge<ThemeOptions>(cometThemeOptions, restPassedOptions);
     return createTheme(themeOptions);
 };