+++ packages/admin/admin/src/error/errordialog/useErrorDialog.ts
@@ -1,7 +1,14 @@
 import * as React from "react";
 
-import { ErrorDialogContext, ErrorDialogContextProps } from "./ErrorDialogContext";
+import { ErrorDialogOptions } from "./ErrorDialog";
+import { errorDialogVar } from "./errorDialogVar";
+export interface UseErrorDialogReturn {
+    showError: (options: ErrorDialogOptions) => void;
+}
 
-export function useErrorDialog(): ErrorDialogContextProps | undefined {
-    return React.useContext(ErrorDialogContext);
+export function useErrorDialog(): UseErrorDialogReturn | undefined {
+    const showError = React.useCallback((options: ErrorDialogOptions) => {
+        errorDialogVar(options);
+    }, []);
+    return { showError };
 }